requirements.txt
torch==1.8.1
torchvision==0.9.1
numpy==1.20.2
pandas==1.2.4
scikit-learn==0.24.2
matplotlib==3.3.4
pillow==8.2.0
dash~=1.21.0
plotly~=5.1.0
seaborn~=0.11.1
scipy~=1.6.2
tensorboard~=2.4.0
Werkzeug~=1.0.1
umap-learn~=0.5.1